backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0

# LangGraph and LangChain
langgraph==0.0.20
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.10

# Vector DB - Pinecone
pinecone==3.0.0

# Resume parsing - Docling
docling>=2.0.0
python-docx>=1.1.2

# HTTP clients
httpx==0.25.2

# Database - SQLite (for development)
aiosqlite==0.19.0
greenlet>=2.0.0

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
email-validator>=2.0.0

# Utilities
python-dateutil==2.8.2

# Open-Source Models (Qwen)
torch>=2.0.0
transformers>=4.40.0
sentence-transformers>=2.7.0
accelerate>=0.27.0
